backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
requests
pydantic
pydantic-settings
python-dotenv
python-dateutil